app.docker.deployment/requirements.txt
pip
Flask
cassandra-driver